+++ app/src/main/java/ch/epfl/sweng/zuluzulu/User/AuthenticatedUser.java
@@ -23,13 +23,13 @@ public class AuthenticatedUser extends User {
     private final String first_names;
     private final String last_names;
     // All followed ids of Associations, Chats and Events
-    private List<Integer> fav_assos;
+    private List<Long> fav_assos;
     private List<Integer> followed_chats;
     private List<Integer> followed_events;
     // TODO add argument to constructor and store liked_event in firebase
     private Set<Integer> liked_events = new HashSet<>();
 
-    protected AuthenticatedUser(String sciper, String gaspar, String email, String section, String semester, String first_names, String last_names, List<Integer> fav_assos, List<Integer> followed_events, List<Integer> followed_chats) {
+    protected AuthenticatedUser(String sciper, String gaspar, String email, String section, String semester, String first_names, String last_names, List<Long> fav_assos, List<Integer> followed_events, List<Integer> followed_chats) {
         firestore_path = "users_info/" + sciper;
         this.sciper = sciper;
         this.gaspar = gaspar;
@@ -69,7 +69,7 @@ public class AuthenticatedUser extends User {
     }
 
     public void removeFavAssociation(Association asso) {
-        fav_assos.remove((Integer) asso.getId());
+        fav_assos.remove(asso.getId());
         Utils.removeIdFromList(firestore_path, "fav_assos", asso.getId());
     }
 
@@ -77,7 +77,7 @@ public class AuthenticatedUser extends User {
         this.followed_chats = followed_chats;
     }
 
-    public void setFavAssos(List<Integer> fav_assos) {
+    public void setFavAssos(List<Long> fav_assos) {
         this.fav_assos = fav_assos;
     }
